// src/delivery/enum/telegram.enum.ts
export enum TelegramChatType {
  GROUP = 'group',
  SUPERGROUP = 'supergroup',
  PRIVATE = 'private',
}

export enum AdminTelebotAction {
  ADD = 'add',
  REMOVE = 'remove',
}

export enum GroupTelebotAction {
  ADD = 'add',
  REMOVE = 'remove',
}

export enum TelegramMenuAction {
  CHECK = 'check',
  UNCHECK = 'uncheck',
  CHOOSE = 'choose',
  NO_ACTION = 'noaction',
  BACK = 'back',
  GO_PAGE = 'gopage',
  ALL_PAGE = 'allpage',
  CHOICE = 'choice',
  TEXT = 'textmenu',
}

export enum DocumentMimeType {
  XML = 'application/xml',
  XML2 = 'text/xml',
  EXCEL = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
}